SSE-8 Rename Heartbeat to Tether
- Besides the refactoring to perform the rename, there is some
re-packaging to distinguish "events" from "keepalive".
