+++ src/main/java/com/clueride/sse/event/answer/AnswerSummaryWebService.java
@@ -13,50 +13,72 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *
- * Created by jett on 2/24/19.
+ * Created by jett on 3/6/19.
  */
-package com.clueride.sse.heartbeat;
+package com.clueride.sse.event.answer;
 
 import java.lang.invoke.MethodHandles;
 
 import javax.inject.Singleton;
+import javax.ws.rs.Consumes;
 import javax.ws.rs.GET;
+import javax.ws.rs.POST;
 import javax.ws.rs.Path;
+import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
 import javax.ws.rs.QueryParam;
+import javax.ws.rs.core.MediaType;
 
 import org.apache.log4j.Logger;
 import org.glassfish.jersey.media.sse.EventOutput;
 import org.glassfish.jersey.media.sse.SseBroadcaster;
 import org.glassfish.jersey.media.sse.SseFeature;
 
+import com.clueride.sse.common.CommonChannelService;
+import com.clueride.sse.common.CommonChannelServiceImpl;
 import com.clueride.sse.common.ServerSentEventChannel;
 
 /**
- * Broadcasts a Heartbeat to subscribers to verify a connection
- * with the server is still alive.
- *
- * For subscribers providing an Outing ID, the heartbeat message will
- * include tethered position for that Outing.
+ * REST API for subscribing to and broadcasting to the Puzzle Events.
  */
 @Singleton
-@Path("heartbeat")
-public class HeartbeatWebService {
+@Path("answer-summary")
+public class AnswerSummaryWebService {
     private static Logger LOGGER = Logger.getLogger(MethodHandles.lookup().lookupClass());
 
-    private HeartbeatService heartbeatService = new HeartbeatServiceImpl();
+    private AnswerSummaryService answerSummaryService = new AnswerSummaryServiceImpl();
+
+    private CommonChannelService commonChannelService = new CommonChannelServiceImpl();
 
     @GET
     @Produces(SseFeature.SERVER_SENT_EVENTS)
-    public EventOutput subscribeToHeartbeat(
+    @Path("{outingId}")
+    public EventOutput subscribeToOutingIdChannel(
+            @PathParam("outingId") Integer outingId,
             @QueryParam("r") final String requestId
     ) {
-        LOGGER.debug("Heartbeat subscribe - requestId: " + requestId);
-        ServerSentEventChannel channel = heartbeatService.openChannelResources(null);
+        LOGGER.debug("Outing ID: " + outingId);
+
+        ServerSentEventChannel channel = commonChannelService.getEventChannel(
+                outingId
+        );
         SseBroadcaster broadcaster = channel.getBroadcaster();
+
         EventOutput eventOutput = new EventOutput();
         broadcaster.add(eventOutput);
         return eventOutput;
+    }
+
+    @POST
+    @Path("{outingId}")
+    @Produces(MediaType.TEXT_PLAIN)
+    @Consumes(MediaType.TEXT_PLAIN)
+    public String broadcastMessage(
+            String answerSummary,
+            @PathParam("outingId") Integer outingId
+    ) {
+        String messageId = answerSummaryService.broadcastMessage(outingId, answerSummary);
+        return "Message ID '" + messageId + "' has been broadcast.";
     }
 
 }
