+++ src/main/java/com/clueride/sse/event/game/README.md
