+++ src/main/resources/log4j.properties
@@ -5,4 +5,4 @@ log4j.appender.stdout.layout=org.apache.log4j.PatternLayout
 log4j.appender.stdout.layout.ConversionPattern=%d [%c] %-5p - %m%n
 
 log4j.logger.com.clueride.sse.CORSFilter=INFO
-log4j.logger.com.clueride.sse.heartbeat=INFO
+log4j.logger.com.clueride.sse.event.tether=INFO
